Extract cucumber configuration to separate class
