+++ cab-app-passenger-service/src/integration-test/java/by/arvisit/cabapp/passengerservice/component/CucumberCTConfiguration.java
@@ -1,0 +1,35 @@
+package by.arvisit.cabapp.passengerservice.component;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.jdbc.Sql;
+import org.springframework.test.context.jdbc.SqlGroup;
+
+import io.cucumber.java.Before;
+import io.cucumber.java.BeforeAll;
+import io.cucumber.spring.CucumberContextConfiguration;
+import io.restassured.RestAssured;
+
+@CucumberContextConfiguration
+@ActiveProfiles("itest")
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@SqlGroup({
+        @Sql(scripts = "classpath:sql/add-passengers.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD),
+        @Sql(scripts = "classpath:sql/delete-passengers.sql", executionPhase = Sql.ExecutionPhase.AFTER_TEST_METHOD) })
+public class CucumberCTConfiguration {
+
+    @LocalServerPort
+    private int serverPort;
+
+    @BeforeAll
+    public static void setUpDB() {
+        System.setProperty("spring.datasource.url", "jdbc:tc:postgresql:15-alpine:///");
+    }
+
+    @Before
+    public void setUpPort() {
+        RestAssured.port = serverPort;
+    }
+}
